moved i18n to src and it's usage
