+++ src/components/assets/asset-card.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import type { Asset } from '../@/lib/types'
+import type { Asset } from '@/lib/types'
 import { Card, CardContent } from '../ui/card'
 import { Badge } from '../ui/badge'
 import { Button } from '../ui/button'
@@ -11,9 +11,9 @@ import {
   DropdownMenuTrigger,
 } from '../ui/dropdown-menu'
 import { Pen } from 'lucide-react'
-import { getStorageUrl } from '../@/lib/utils'
+import { getStorageUrl } from '@/lib/utils'
 import { getAssetIcon, getAssetTypeLabel } from './asset-utils'
-import { Icons } from '../@/lib/constances'
+import { Icons } from '@/lib/constances'
 import { useTranslation } from 'react-i18next'
 
 interface AssetCardProps {
