+++ src/components/assets/project-selector.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState, useEffect } from 'react'
-import type { Project } from '../@/lib/types'
+import type { Project } from '@/lib/types'
 import { Button } from '../ui/button'
 import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover'
 import {
@@ -12,8 +12,8 @@ import {
   CommandItem,
   CommandList,
 } from '../ui/command'
-import { cn } from '../@/lib/utils'
-import { Icons } from '../@/lib/constances'
+import { cn } from '@/lib/utils'
+import { Icons } from '@/lib/constances'
 import { useTranslation } from 'react-i18next'
 
 interface ProjectSelectorProps {
