+++ src/components/assets/rename-dialog.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState, useEffect } from 'react'
-import { Asset } from '../@/lib/types'
+import { Asset } from '@/lib/types'
 import { Button } from '../ui/button'
 import { Input } from '../ui/input'
 import { ModalWrapper } from '../ui/modal-wrapper'
