+++ src/components/dashboard/RecentItems.tsx
@@ -1,17 +1,17 @@
 'use client'
 
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '../@/components/ui/tabs'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '../@/components/ui/card'
-import { Button } from '../@/components/ui/button'
-import { Badge } from '../@/components/ui/badge'
-import { Separator } from '../@/components/ui/separator'
-import { cn } from '../@/lib/utils'
-import { Icons } from '../@/lib/constances'
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
+import { Button } from '@/components/ui/button'
+import { Badge } from '@/components/ui/badge'
+import { Separator } from '@/components/ui/separator'
+import { cn } from '@/lib/utils'
+import { Icons } from '@/lib/constances'
 import { useTranslation } from 'react-i18next'
 import { useMobile } from '../../hooks/use-mobile'
 import { CheckCircle2, PenTool } from 'lucide-react'
-import { getAssetTypeLabel, getStorageUrl } from '../@/lib/utils'
-import type { Project, Label, Asset } from '../@/lib/types'
+import { getAssetTypeLabel, getStorageUrl } from '@/lib/utils'
+import type { Project, Label, Asset } from '@/lib/types'
 import type { NavigateFunction } from 'react-router-dom'
 
 import { useState, useEffect } from 'react'
