+++ src/components/ui/command.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react'
 import { Command as CommandPrimitive } from 'cmdk'
 
-import { cn } from '../@/lib/utils'
+import { cn } from '@/lib/utils'
 import {
   Dialog,
   DialogContent,
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from '../@/components/ui/dialog'
-import { Icons } from '../@/lib/constances'
+} from '@/components/ui/dialog'
+import { Icons } from '@/lib/constances'
 
 function Command({ className, ...props }: React.ComponentProps<typeof CommandPrimitive>) {
   return (
