+++ src/components/ui/dialog.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import * as DialogPrimitive from '@radix-ui/react-dialog'
 
-import { cn } from '../@/lib/utils'
-import { Icons } from '../@/lib/constances'
+import { cn } from '@/lib/utils'
+import { Icons } from '@/lib/constances'
 
 function Dialog({ ...props }: React.ComponentProps<typeof DialogPrimitive.Root>) {
   return <DialogPrimitive.Root data-slot="dialog" {...props} />
