+++ src/components/ui/modal-wrapper.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import type React from 'react'
-import { cn } from '../@/lib/utils'
+import { cn } from '@/lib/utils'
 
 interface ModalWrapperProps {
   isOpen: boolean
