+++ src/components/ui/slider.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import * as SliderPrimitive from '@radix-ui/react-slider'
 
-import { cn } from '../@/lib/utils'
+import { cn } from '@/lib/utils'
 
 function Slider({
   className,
