+++ src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import '../i18n.tsx'
-import './index.css'
-import App from './App.tsx'
+import '@/i18n.tsx'
+import '@/index.css'
+import App from '@/App.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
