+++ src/pages/parts/side-panel.tsx
@@ -1,18 +1,18 @@
 'use client'
 
-import { useAuth } from '../@/context/AuthContext'
+import { useAuth } from '@/context/AuthContext'
 import { Link, useLocation } from 'react-router-dom'
-import { Button } from '../@/components/ui/button'
-import { Icons } from '../@/lib/constances'
-import { cn } from '../@/lib/utils'
+import { Button } from '@/components/ui/button'
+import { Icons } from '@/lib/constances'
+import { cn } from '@/lib/utils'
 import { useMobile } from '../../hooks/use-mobile'
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from '../@/components/ui/tooltip'
-import { Avatar, AvatarFallback, AvatarImage } from '../@/components/ui/avatar'
+} from '@/components/ui/tooltip'
+import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import { useTranslation } from 'react-i18next'
 
 type SidePanelProps = {
