+++ tsconfig.app.json
@@ -33,6 +33,5 @@
   },
   "include": [
     "src",
-    "i18n.tsx"
   ]
 }
